Move proto directory back to top level. (#21)
